+++ PointCloudRetouch/PointSetPreprocessor.h
@@ -11,7 +11,7 @@ namespace hiveObliquePhotography
 			~CPointSetPreprocessor() = default;
 			
 			void cullByDepth(std::vector<pcl::index_t>& vioPointSet, const Eigen::Matrix4d& vPvMatrix, const Eigen::Vector3d& vViewPos);
-			void cullBySdf(std::vector<pcl::index_t>& vioPointSet, const Eigen::Matrix4d& vPvMatrix, const std::function<float(Eigen::Vector2f)>& vSignedDistanceFunc);
+			void cullBySdf(std::vector<pcl::index_t>& vioPointSet, const Eigen::Matrix4d& vPvMatrix, const std::function<double(Eigen::Vector2d)>& vSignedDistanceFunc);
 		private:
 			std::pair<Eigen::Vector2d, Eigen::Vector2d> __computeBoundingBoxOnNdc(const std::vector<pcl::index_t>& vPointSet, const Eigen::Matrix4d& vPvMatrix);
